Minor efficiency improvement in findBestMethod.
Since closeness can't be both incremented AND decremented in the same pass, an else was added. (very minor, but worthwile)

Also the comments were corrected, subclass and superclass often get mixed up.

This is a special update for CVS users, as I will wait for something more significant to update the zipped codebase file :-)