Sweeping optimization! Functionally and architecturally identical. Further integrated the static Remote.invoke functionality throught the architecture. A significant benefit is that now a proxy can be any object, and doesn't even have to implement Invoke. This will make integration into legacy applications significantly easier. Many classes affected, but the result is greater flexibility, and even reduced code size!
Affected classes:
ProxyLoader: now accepts any object, rather than those implementing Invoke.
Client: now accepts any type of object from a remote item.
ItemServer now accepts both proxies and items as objects, in fact there was a small bugfix, if an item was an object and not wrapped in a MonitorItem, it would not have been configured properly, this is now fixed.
BaseProxy now is simply an object.
Remote the static invoke method has been further generalized, for its increased use. It now checks if the item to be invoked implements invoke, if so, the invocation is passed on to its method.
MonitorItem now uses the improved static Remote.invoke method.

Whew! A lot of refining, but ultimately a better, smaller more flexible implementation. Time for a beer...