A cleanup release, fully backward compatible.
Files affected:

Client.java: Uses Remote.Invoke(object, method, args) instead of invoke(method, args) which is the preferred technique, because it works for all objects, not just those implementing Invoke.

Remote.java: Removed use of NoSecurityManager in the static main method, this created a potential security hole, if it was not explicitly disabled, so now it must be explicitly enabled. It also caused the class to be included in the proxy and server jar files, even though it wasn't used.

MonitorItem.java: A nice improvement whereby the monitor 'peeks' into the request and responses, if they are wrapped in a MarshalledObject. It provides more information to assist in debugging.

BaseItem.java: It used to reference the frame member class in Client.java. This caused all the client classes to be included in the server jar, even when they weren't needed. Also, the automatic framing of graphical proxies was removed, as it could overload a client.

Now the make3.bat will generate the smallest possible jar files! :-)