Update to indicate a new codebase release!
It contains a slick generalization to BaseItem, which allows the setProxy method to accept MarshalledObjects of proxy items. This allows remote items to load proxies into other remote items, this was not possible before. Also, this is nicely and fully backward compatible with the previous edition. It just keeps getting better and better!

And a  minor update to ItemServer, to make use of the java.rmi.server.useCodebaseOnly to further prevent loading of remote codebases, unless the acceptProxies method is called. It is minor, but it may improve security. Either way, the ItemServer works identically as before.