Two cool new updates:
A new class AuditorItem is added to the extra package.
It cleanly illustrates adding the Decorator pattern to the framework.
It allows arguments going into, and results returned by an object to be
examined and even modified. Ideal for debug, and application security.

CodebaseServer is updated with a cool new hack courtesy of Laird Nelson!
This allows the server's jar file to be named freely, while still protecting it from
being uploaded by an overly curious client. Prior to this the server's jar had
to be named "server.jar" to receive this protection. All-in-all a very neat
improvement!