Added an exciting new class to gnu.cajo.utils.extra
It is called Registry, it is a server, which broadcasts its startup, and allows
other VMs to register an object with it, either statically or dynamically. A VM
can also request a hashtable of its current registries.

The broadcast at startup allows registries to spontaneously reference each
other. It is a good general foundation for the WWVM.