Two super-cool updates!
First, on re-reading the RMI API docs, I noticed they recommend that custom
SocketFactories implement the equals and hashCode methods. Upon further
investigation, I found that doing this would enable the socket connection
pooling feature of RMI. Well, it's done, and wow; do things run faster! :-)

Second, the Multicast class was updated to give it complete functional
symmetry with ItemServer. Now objects passed to announce will have their
startThread and setProxy methods invoked, and announce also supports
a proxy argument. This allows you to use ItemServer or Multicast
interchangeably. (Don't know why I didn't think of this sooner!)

PS As a super-minor change, the announcement in the example was
commented out. Since the example object is monitored, when the announce
is called after the bind, an exception is generated. This is now to be
expected, and everyting still works correctly, but it did not look pretty on the
console. If you want to see it, feel free to un-comment it.