Today's update is all about timing, and patience:
CodebaseServer now has a timeout for clients. It requires them to make their
request for data within 1500 milliseconds. This should be more than enough time for
applications. The intent is to prevent a locked-up client from also locking up the
server's socket.

Also added setSoTimeout to the RemoteClientSocketFactory. This one times
out if a server has not responded to an invocation after 60 seconds. If a
server needs more time than that, it really should use a callback mechanism.
The intent is to provide a speedy detection of a server performance issue.

Together, these remove two opportunities for servers and clients to 'hang'