several nice updates!
CodebaseServer has been modified to work correctly when not contained in a jar file.
A library.bat file is now included to match the functionality of the library target
in the ant script.
Remote SocketFactories have been made more logically symmetrical, also a
serialVersionUID has been added to RCSF. This corrects a conflict between
Sun and IBM compiler generated UIDs.
This was also done for HashedProxy, ZippedProxy, and ProxyLoader.