two refinements, sort-of linked:
now when a fact is asserted, all rules listening for subclasses of the fact will
be nofitied as well.

when a new rule is added, the memory will be searched. any fact matching
the key will be invoked on the rule, as well as any facts which are superclasses
of the key.