Important update to the invoke method of gnu.cajo.invoke.Remote; to further enhance dynamic dispatch. For Example:
A method foo(Object x) will be invoked if someone invokes with a null argument. (only assuming a method foo() is not already defined)

A method foo(Object x) will be invoked if someone invokes with an array of Strings for example, (only assuming a method foo(String[]) is not already defined)

Prior to this release; These would not work correctly. The result is much better coverage of edge cases, and everything works even more intuitively.

Happy Day! :)