update to address a very important issue, lifetime of remote object references
normally when an object is made remotely accessible, it is for the life of the
server. typically, this is what one would expect.

sometimes however, a "special-case" reference to an object specifically
created for short-term use, is provided to one or more clients. Normally, the
server would not even retain a reference to the remote wrapper itself

however, in this second case, the remote reference will not be garbage
collected when the clients are finished with it, this will create a *memory leak*

To fix this, a new instance method has been added to Remote, called
clientScope. it means literally that, the lifetime of this remote reference is
determined by the client, or clients, rather than the server. It returns a
reference to itself, so that in cases where the otherwise leaky construct:

return new Remote(tempObject);

is replaced by the proper implementation:

return new Remote(tempObject).clientScope();

This addresses a very subtle and tricky potential problem